make torchao test discovery pass in fbcode
Summary:
Running `buck test @//mode/dev-nosan fbcode//pytorch/ao/test:torchao_tests` used
to work but stopped working sometime in the past ~months due to test
discovery being broken. This diff fixes test discovery and adds a blocklist
to all of the broken tests, so we can at least run the passing tests easily.

Note that for now we have some tests using `pytest` which is handled in a custom
way. In a future PR it would be great to make torchao use a single testing
framework.

Reviewed By: jerryzh168

Differential Revision: D58422127