Add semi-structured sparse perf/accuracy benchmarks for SAM
This PR adds in segment-anything-fast evaluation to torchao, and also
adds benchmarks for int8 quantization + 2:4 sparsity.

With this we can run combined perf/accuracy benchmarks for
segment-anything. This should give us a starting point for the relative
perf vs relative acc graph for PTC.

This just copies over the evaluation scripts. Eventually I think we
should move over the modeling code too, but plan to do that in a
subsequent PR.

ghstack-source-id: d0cc5863b4426e438e17b42669e75901862cc067
Pull Request resolved: https://github.com/pytorch-labs/ao/pull/464