Fix regression tests
Summary:

Torch 2.2 is compiled with numpy 1.x, but when we `pip install -r
requirements-dev.txt` we download a higher version of numpy (2.0)

This causes an error with the .numpy() calls and importing torch in
general. I don't think we want to pin the versions in
requirements-dev.txt, so instead I added a pin to the numpy version in
the specific torch spec, so it'll only run for 2.2.

PT 2.3+ support numpy 2.0+ which is why those test don't fail.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: