add roofline estimation of float8 gemm + overhead
Summary:

This PR adds a script to help estimate when it is expected
to be faster to convert a `torch.nn.Linear` to float8, versus
leaving it in bfloat16, for compute bound training use cases.

How we do this:
1. for bf16 and fp8 gemm time, use either benchmarks or roofline based
   on a percentage of peak compute bandwidth
2. for float8 overhead, enumerate the worst case of reads and writes performed
   by torch.compile, and convert to time by assuming we can achieve a
   percentae of peak memory bandwidth
3. compare (a) bf16_gemm_time and (b) fp8_gemm_time + fp8_overhead_time,
   and calculate the expected speedup.

Note that currently the gemm benchmarks require running a separate
script, as documented in the argument descriptions.  We can make this
nicer at a future time.

Test Plan:

```
python benchmarks/float8/float8_roofline.py \
    --outfile ~/local/tmp/20240813_del_del_dyn_bench_roofline.csv \
    --gemm_benchmarks_file ~/local/tmp/20240813_gemm_gpu_time_sweep_9_16.csv \
    --gemm_time_strategy benchmarks --model_torch_compile_limitations True \
    --scaling_type_input delayed --scaling_type_weight delayed \
    --scaling_type_grad_output dynamic | with-proxy gh gist create
- Creating gist...
✓ Created secret gist
https://gist.github.com/vkuzo/e3c4b274493140a7423c85d27863663a
```

Reviewers:

Subscribers:

Tasks:

Tags: