Setup dirsync for torchao experimental (#719)
Summary:
Pull Request resolved: https://github.com/pytorch/ao/pull/719

In order to build and test kernels internally for mac, android and ios we need
to add build support in xplat. Which means we need to setup mirroring between
fbcode and xplat. For now setting this up only for experimental folder, pending
larger decision on whether to sync the two folder entirely.

Reviewed By: mzlee

Differential Revision: D61480451