[torchao] Remove unneeded FFHT files
We are just using FFHT to generate a kernel, so we don't need
the Python wrapper. Likewise, we don't need the full README, so I
replaced it with an explanation of what's going on and the required
copyright notice. I'm not sure what the recommended practice is for
licensing/copyright notice on these files going forward though; please advise?

Differential Revision: [D60194971](https://our.internmc.facebook.com/intern/diff/D60194971/)

ghstack-source-id: 241394062
Pull Request resolved: https://github.com/pytorch/ao/pull/830