[torchao] FFHT: ARM NEON port
Patch the code generator to be capable of generating NEON
code and leave it configured to do that since we already have the
checked-in generated AVX and SSE code (and no immediate plans to use
it). Generated code size was a potential issue so I also patched the
generator to 1) reuse generated code for previous smaller sizes where
applicable and 2) choose the smallest code that isn't more than 10%
slower than the very fastest code.

Differential Revision: [D60194970](https://our.internmc.facebook.com/intern/diff/D60194970/)

[ghstack-poisoned]