[not for land] hardcoded two-stage reduction for max(abs(tensor))
Summary:

Based on investigations around
https://github.com/pytorch/pytorch/issues/128063, changing the user code
to do the tensorwise max in two steps helps the compiler find better
fusion opportunities.

TBD on if we want to land this or have torch.compile do this
automatically, but for now putting up a PR to make this easier to
explore/benchmark.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: