fix local float8 tests on H100
Summary:

float8 tests were failing my machine. I bisected the failure to
https://github.com/pytorch/ao/pull/1344. Further investigation found
that that PR was fine, but the tolerance for one of the tests was too
tight, adding a test changed the random seed of the data, and things
started failing.

Switching to SQNR for a more robust measurement.

Test Plan: CI

Reviewers:

Subscribers:

Tasks:

Tags: