[cleanup][2/x] split float8 mm by delayed vs dynamic
Summary:

Before this PR, the float8 mm logic was split by axiswise vs tensorwise.
After this PR, the float8 mm logic is split by dynamic vs non-dynamic scaling.

Motivation: there is more and more evidence that dynamic scaling will
be common to the most important lowp recipes. This PR is a step on the way
to making the dynamic scaling logic be simpler and easier to understand
in `torchao.float8`.

There are a lot of other simplifications to do, but stopping here to
keep the PR small.  This is a pure refactor without any logic changes.

Test Plan:

```
./test/float8/test_everything.sh
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 90b042fc76549c88a46abf06945ad166ea403e5c
ghstack-comment-id: 2564049757
Pull Request resolved: https://github.com/pytorch/ao/pull/1461