mx: add ceil and RNE rounding modes to the cast from fp32 to e8m0
Summary:

Why we want this: in the newly released cuBLAS 12.8 documentation, the
RNE rounding mode is used for the cast to e8m0. We want to properly
emulate this cast.

This is a copy-pasta of https://github.com/pytorch/ao/pull/516 with the
modifications being keeping FLOOR the default more, and removing e3m0, credit
to NicoleMayer for the original code. I don't have a way to check bitwise
equivalency with the most recent cuBLAS version yet, but will come back
and add tests when I do.

Test Plan:

```
pytest test/prototype/mx_formats/ -s -x
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-comment-id: 2613597310

ghstack-source-id: a160d4c9320301e0d3d62128ad386b7e0c985d59
ghstack-comment-id: 2623497069
Pull Request resolved: https://github.com/pytorch/ao/pull/1643