delete delayed scaling from torchao.float8
Summary:

Previously we added a deprecation warning for delayed scaling in v0.9.0.

Now that the branch cut for v0.9.0 has passed, we are deleting delayed
scaling from the codebase to simplify the code.

Motivation:
1. delayed scaling adds a lot of complexity
2. we still don't have any customers for delayed scaling, since people
   prefer to optimize for higher accuracy from dynamic scaling

If a need for delayed scaling arises in the future, we can bring it back
to a prototype folder in a way that is decoupled from the dynamic
scaling code.

Test Plan:

```
./test/float8/test_everything.sh
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 8d8b801dff49901c211389214eae0c3e79774768
ghstack-comment-id: 2674970350
Pull Request resolved: https://github.com/pytorch/ao/pull/1753