Write weight packing/unpacking functions for universal kernels (#1921)
Summary:
Pull Request resolved: https://github.com/pytorch/ao/pull/1921

This diff implements unpacking functions for the existing universal kernels.

To strealine the changes across the 3 tile sizes, I also rewrote the packing functions to be parametrized by nr, kr, sr.  In this way, I only needed to rewrite the unpacking function once.

New tests are added to show that when we pack and then unpack, we recover the weight qvals, scales, zeros, and bias.

This diff is most of the work for implementing an embedding kernel that understands the universal linear packing format.

The remaining work is:

* Write unpack to accept n_idx as an arg, rather than compute all columns.
* Use the unpack weight function in the embedding kernel, followed by a dequant op.

Reviewed By: digantdesai

Differential Revision: D71163696