Reintroduce has_weight_zeros as a template param (#1991)
Summary:
Pull Request resolved: https://github.com/pytorch/ao/pull/1991

When introducing bias, we made has_weight_zeros a regular argument instead of a template arg.

I've noticed this has decreased perf, so we re-introduce has_weight_zeros as a template arg.  I keep has_bias and has_clamp as regular args because they are only applied in the outer most loop, and we want to reduce code bloat.

Note that all of this happens during internal dispatch, and the user-level python op has no change in API.

Reviewed By: drisspg

Differential Revision: D71503133