create staticmethod for quantizing weights of QATLinear and QATEmbedding (#2079)
Summary:

For saving the quantized weights, we have been using adhoc notebooks with copy-pasted code from the convert method.
This had been a source of numerical discrepancies. To avoid this issue, this diff adds separates the weight quantization logic in to a separate staticmethods so that we can reuse it.

Reviewed By: jerryzh168

Differential Revision: D73201409