ROCm mxfp4 Skips (#2209)
Add skip_if_rocm decorator to test_inference_subclass in test_mx_linear.py

This change introduces the skip_if_rocm decorator to the test_inference_subclass function, ensuring that the test is skipped on ROCm platforms due to the requirement for gfx950. This enhances the robustness of the test suite by preventing unnecessary failures on unsupported hardware.