Add selective weight loading decode kernel for activation sparsity
Summary:

This PR adds in a kernel to accelerate $$xW^T$$ when x is sparse and we
are memory bound.

The idea here is that we can avoid loading the columns of $W$ that
correspond to the zero elements of $x$.

This lets us accelerate activation sparsity for bs=1 decode use cases.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: