[sparse] remove superblock (#2381)
Removing superblock folder from prototype, as the superblock subclass and kernel have already previously been promoted out of prototype to torchao.sparsity.

All that remains are custom training loops which is no longer used as it's for ViTs and the current block sparse work we are investigating is for LLMs.