remove torchao dependency from torchao build script (#2383)
Summary:

We should not use torchao to build torchao. Before this PR,
I got stuck in a loop where I had a broken install (for some other
reason) and could not uninstall torchao with this setup script,
because the script required a working torchao to run.

Commenting out the offending code in this PR, leaving for a
future person to actually fix it.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: