fix float8 training TP+SP integration tests
Summary:

These tests do not run in CI, and they broke some time ago. The issue
was that each tensor was created on "cuda:0" instead of using the local
rank. For now, fixing by manually specifying the rank. I feel like there
is probably a better way to do this as the rank is supposed to be set
automatically, but leaving that for a future PR.

We should add to CI in the future, saving that for a future PR.

Test Plan:

```bash
./test/float8/test_dtensor.sh
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 39bd880290d3f976c66ed6797b787ee0200e7b82
ghstack-comment-id: 2991778315
Pull Request resolved: https://github.com/pytorch/ao/pull/2414