release notes script: keep not user facing rows (#2875)
Summary:

We should keep the "not user facing" PRs in the draft release notes so
that a human can go through them and verify they are actually not user
facing.  It's still easy to delete all of them from the generated file
in a few seconds.

Test Plan:

Used this for v0.13.0 release notes formatting

Reviewers:

Subscribers:

Tasks:

Tags: