mxtensor: switch to AOBaseTensor dispatch
Summary:

Deletes the custom op dispatch logic in `MXTensor` and switches to the
general one in `AOBaseTensor`.

To enable this, we move the mx ops to the same file as `MXTensor`. This
is to avoid the need for callsites to separately import `mx_ops.py`.

Note that there are a couple of custom mx ops which could instead use
the general implementation, leaving that for future PRs to keep this one
small.

Test Plan:

```
pytest test/prototype/mx_formats -s -x
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 9dd0007cefd8343d56200b5f344b73586272994a
ghstack-comment-id: 3338749802
Pull Request resolved: https://github.com/pytorch/ao/pull/3080