Prevent union-finding cycles for shared qspecs (#3011)
* Prevent union-finding cycles for shared qspecs

Certain graphs caused infinite recursion when unioning nodes to groups
based on shared quantization specs while implicit sharing was enabled.
The problem occurred when `NodeOrEdge` A with its own `QuantizationSpec`
received an edge (in `shared_with_map`) to `EdgeOrNode` B which in turn
had a `SharedQuantizationSpec` pointing back to A.

Remedy this problem by checking if B, from the scenario above, has a
`SharedQuantizationSpec` pointing to A; if that is the case, don't union
them together by letting A point back to B. Avoiding the union/cycle
preserves correctness because the nodes are effectively already united.

* Add test case of implicit sharing with two ops sharing one input

- Add test case of implicit sharing for a model where one input is
  shared between two different ops.
- Add code comments to `_union_if_no_cycle`

* Simplify the model that reproduces the recursion bug

* Avoid forming a cycle by reversing the edge

* Add context of recursion bug in the test case

* Swap order of branches in if statement

---------

Co-authored-by: Martin Lindström <Martin.Lindstroem@arm.com>