Remove old TORCH_VERSION variables (#3146)
* Remove config functions like `int4_weight_only`

**Summary:** As a follow-up to https://github.com/pytorch/ao/pull/2994,
this commit removes all quantization functions that were used
as configs. These functions were deprecated in 0.14.0 and will
be removed in the next release, 0.15.0.

**Test Plan:** CI

[ghstack-poisoned]

* Remove old TORCH_VERSION variables

**Summary:** As a follow-up to https://github.com/pytorch/ao/pull/2719,
which deprecated these variables in 0.13.0, we remove them now in
the next release 0.15.0.

**Test Plan:** CI

[ghstack-poisoned]

* Update base for Update on "Remove old TORCH_VERSION variables"


**Summary:** As a follow-up to https://github.com/pytorch/ao/pull/2719,
which deprecated these variables in 0.13.0, we remove them now in
the next release 0.15.0.

**Test Plan:** CI

[ghstack-poisoned]

* Update base for Update on "Remove old TORCH_VERSION variables"


**Summary:** As a follow-up to https://github.com/pytorch/ao/pull/2719,
which deprecated these variables in 0.13.0, we remove them now in
the next release 0.15.0.

**Test Plan:** CI

[ghstack-poisoned]