[pt2e] Fix QAT annotations for special qspecs (#3337)
**Summary:** In pt2e QAT, we first annotate the nodes to be
quantized and then perform the pattern replacement (i.e. QAT
fusion) in the prepare step. After this pattern replacement,
old nodes are replaced with new nodes, and any references to
the old nodes must also be updated to refer to the new nodes.

This commit fixes a bug where, for special qspecs like the
`SharedQuantizationSpec` and `DerivedQuantizationSpec`, we only
update the values of a node's `input_qspec_map`, not the keys.
As a result, the keys still refer to the old nodes that do
not exist anymore after the QAT fusion.

**Test Plan:**
```
python test/quantization/pt2e/test_quantize_pt2e_qat.py -k test_qat_shared_qspec
```