2/x mx cleanup: remove pack_fp6
Summary:

This functionality is useful for fp6 research, but it is going to need a
new home as we make the mxfp8 and mxfp4 paths production ready. Removing
it for now, and we are happy to work with collaborators to add it back
in a way which is isolated from the near-term production path, as
needed.

Test Plan:

```
pytest test/prototype/mx_formats -s -x
```

Reviewers:

Subscribers:

Tasks:

Tags:
ghstack-source-id: 620b71fd740a07e4bd10284e894e7bd95b1053a6
ghstack-comment-id: 3571962192
Pull-Request: https://github.com/pytorch/ao/pull/3382